requirements.txt
python-dateutil~=2.8.2 ; python_version < "3.11"